Simplified the scheduler and MAC. DL and UL slots are not scheduled in different subframes now.
Instead we schedule DL and UL slots in the same subframe, which applies
to a future subframe some number of TTIs in the future, configured by
setting the "L1CtrlLatency" common parameter.  The new code needs
extensive cleaning up and documentation still.

+++ src/mmwave/model/mmwave-ue-mac.h
@@ -118,7 +118,7 @@ private:
 	};
 
 	std::map <uint8_t, LcInfo> m_lcInfoMap;
-	uint8_t m_rnti;
+	uint16_t m_rnti;
 
 	bool m_waitingForRaResponse;
 	static uint8_t g_raPreambleId;
